Clean up unused code and files Rename main flask code folder to app Rename app.py to InitAPI.py
