+++ getvite_api/Dockerfile
@@ -1,6 +1,6 @@
 FROM python:3.7-alpine
 WORKDIR /api
-ENV FLASK_APP app.py
+ENV FLASK_APP InitAPI.py
 ENV FLASK_ENV=development
 ENV FLASK_RUN_HOST 0.0.0.0
 RUN apk add --no-cache gcc musl-dev linux-headers
